[Feat] #9 ReadShortcutView 전체 구성
ReadShortcutView 전체 구성
