--- HappyAnding/HappyAnding/Views/ShortcutDetailViews/ReadShortcutView.swift
@@ -1,20 +1,0 @@
-//
-//  ReadShortcutView.swift
-//  HappyAnding
-//
-//  Created by 이지원 on 2022/10/19.
-//
-
-import SwiftUI
-
-struct ReadShortcutView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct ReadShortcutView_Previews: PreviewProvider {
-    static var previews: some View {
-        ReadShortcutView()
-    }
-}
