+++ HappyAnding/HappyAnding/Views/ShortcutDetailViews/ReadShortcutView/ReadShortcutContentView.swift
@@ -1,0 +1,22 @@
+//
+//  ReadShortcutContentView.swift
+//  HappyAnding
+//
+//  Created by KiWoong Hong on 2022/10/24.
+//
+
+import SwiftUI
+
+struct ReadShortcutContentView: View {
+    var body: some View {
+        ScrollView {
+            Text("ReadShortcutContentView")
+        }
+    }
+}
+
+struct ReadShortcutContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ReadShortcutContentView()
+    }
+}
