+++ HappyAnding/HappyAnding/Views/ShortcutDetailViews/ReadShortcutView/ReadShortcutHeaderView.swift
@@ -1,0 +1,20 @@
+//
+//  ReadShortcutHeaderView.swift
+//  HappyAnding
+//
+//  Created by KiWoong Hong on 2022/10/24.
+//
+
+import SwiftUI
+
+struct ReadShortcutHeaderView: View {
+    var body: some View {
+        Text("ReadShortcutHeaderView")
+    }
+}
+
+struct ReadShortcutHeaderView_Previews: PreviewProvider {
+    static var previews: some View {
+        ReadShortcutHeaderView()
+    }
+}
